Simplistic Chatbot
This script uses Python to function, it's a simplistic chatbot where you need to register/login to use. You'll need to create a "chat_history.txt" and "details.txt" in the same directory as the main.py

The chat_history.txt stores previous conversations in a text file format even after you close the chat. The details.txt does the exact same thing but for user information when users register/login. None of these credentials are encrypted as this is just an experiment/example.